.history/requirements_20260129152929.txt
flask
gunicorn
flask-cors
psycopg2-binary
python-dotenv
